requirements.txt
Django==1.7.7
GDAL==1.9.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
astroid==1.2.0
billiard==3.3.0.18
celery==3.1.15
django-pipeline==1.3.25
django-tastypie==0.11.1
futures==2.1.6
httplib2==0.9
kombu==3.0.23
librabbitmq==1.5.2
logilab-astng==0.24.3
logilab-common==0.62.0
mimeparse==0.1.3
oauth2==1.5.211
oauthlib==0.6.3
psycopg2==2.5.3
pylint==1.3.0
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.7
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
wsgiref==0.1.2
django-require==1.0.6
